Split CurriculumVitaeGetterTest file
