+++ Tests/Entity/CurriculumVitaeGetterFromExampleXMLFileTest.php
@@ -13,7 +13,7 @@ namespace FabienCrassat\CurriculumVitaeBundle\Tests\Entity;
 
 use FabienCrassat\CurriculumVitaeBundle\Entity\CurriculumVitae;
 
-class CurriculumVitaeGetterTest extends \PHPUnit_Framework_TestCase
+class CurriculumVitaeGetterFromExampleXMLFileTest extends \PHPUnit_Framework_TestCase
 {
     private $curriculumVitae;
     private $lang;
@@ -22,69 +22,6 @@ class CurriculumVitaeGetterTest extends \PHPUnit_Framework_TestCase
 
     public function __construct() {
         $this->lang = 'en';
-    }
-
-    public function testGetLookingForAndExperiencesAndHumanFileName() {
-        $this->curriculumVitae = new CurriculumVitae(__DIR__.'/../Resources/data/backbone.xml', $this->lang);
-
-        $result = array();
-        $result = array_merge($result, array('lookingFor' => $this->curriculumVitae->getLookingFor()));
-        $result = array_merge($result, array('experiences' => $this->curriculumVitae->getExperiences()));
-        $result = array_merge($result, array('pdfFile' => $this->curriculumVitae->getHumanFileName()));
-
-        $expected = array(
-            'lookingFor' => array(
-                'experience'   => array(
-                    'date' => 'Date',
-                    'job' => 'The job',
-                    'society' => array(
-                        'name' => 'My Company',
-                        'address' => 'The address of the company',
-                        'siteurl' => 'http://www.MyCompany.com')),
-                'presentation' => 'A presentation'),
-            'experiences' => array(
-                'LastJob' => array(
-                    'date' => 'Date',
-                    'job' => 'The job',
-                    'society' => array(
-                        'name' => 'My Company',
-                        'address' => 'The address of the company',
-                        'siteurl' => 'http://www.MyCompany.com'))),
-            'pdfFile' => 'First Name Last Name - The job'
-        );
-        $this->assertEquals($expected, $result);
-    }
-
-    public function testGetAnchorsWithNoLang() {
-        $this->curriculumVitae = new CurriculumVitae(__DIR__.'/../Resources/data/backbone.xml');
-
-        $anchors = $this->curriculumVitae->getAnchors();
-        if (is_array($anchors)) {
-            $this->assertEquals(array('identity' => array(
-                        'href' => 'identity',
-                        'title' => 'identity'),
-                      'followMe' => array(
-                        'href' => 'followMe',
-                        'title' => 'followMe'),
-                      'experiences' => array(
-                        'href' => 'experiences',
-                        'title' => 'experiences'),
-                      'skills' => array(
-                        'href' => 'skills',
-                        'title' => 'skills'),
-                      'educations' => array(
-                        'href' => 'educations',
-                        'title' => 'educations'),
-                      'languageSkills' => array(
-                        'href' => 'languageSkills',
-                        'title' => 'languageSkills'),
-                      'miscellaneous' => array(
-                        'href' => 'miscellaneous',
-                        'title' => 'miscellaneous')
-                ),
-                $anchors
-            );
-        }
     }
 
     public function testGetIdentityWithEnglishLanguage() {
